+++ docs/whats_new.rst
@@ -27,6 +27,7 @@ Enhancements
 - Adding :class:`braindecode.models.SSTDPN`  (:gh:`790` by `Can Han`_ and `Bruno Aristimunha`_ )
 - Added :class:`braindecode.models.BENDR`  (:gh:`802` by `Bruno Aristimunha`_ )
 - Adding a tutorial for fine-tuning a pre-trained foundation model (:gh:`794` by `Pierre Guetschel`_)
+- Adding artifact removal preprocessing steps, incl. :class:`braindecode.preprocessing.EEGPrep` (:gh:`796` by `Christian Kothe`_)
 - Including huggingface hub mixin to all models (:gh:`795` by `Kuntal Kokate`_ )
 - Better inheritance between the different dataset classes (:gh:`806` by `Pierre Guetschel`_ )
 
@@ -412,4 +413,5 @@ Authors
 .. _Hamdi Altaheri: https://github.com/Altaheri
 .. _José Mauricio: https://github.com/josemaurici
 .. _Can Han: https://github.com/hancan16
+.. _Christian Kothe: https://github.com/chkothe
 .. _Kuntal Kokate: https://github.com/Kkuntal990